maybe last commit!

+++ mpproj/src/model/PremiumPackagesModel.java
@@ -2,11 +2,11 @@ package model;
 public enum PremiumPackagesModel {
     FIRST(5, 30), SECOND(9, 60), THIRD(14, 180);
 
-    private final int cash;
+    private final double cash;
     private final int days;
 
 
-    PremiumPackagesModel(int cash, int days) {
+    PremiumPackagesModel(double cash, int days) {
         this.cash = cash;
         this.days = days;
     }
@@ -15,7 +15,7 @@ public enum PremiumPackagesModel {
         return this.days;
     }
 
-    public int getCash() {
+    public double getCash() {
         return this.cash;
     }
 }
